# frontend/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "default": {},
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "EvidenceRef": {
        "description": "Where a fact was read: page ``page_no`` (1-based) of ``document_id``, highlighting the\n``span_ids`` (GET /documents/{document_id}/pages/{page_no} returns their boxes).",
        "properties": {
          "doc_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Type"
          },
          "document_id": {
            "title": "Document Id",
            "type": "integer"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "page_no": {
            "title": "Page No",
            "type": "integer"
          },
          "span_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Span Ids",
            "type": "array"
          }
        },
        "required": [
          "document_id",
          "page_no",
          "span_ids",
          "filename",
          "doc_type"
        ],
        "title": "EvidenceRef",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JsonValue": {},
...
      "LoginIn": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginIn",
        "type": "object"
      },
      "Me": {
        "properties": {
          "auth_mode": {
            "title": "Auth Mode",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "permissions": {
            "description": "What the roles allow; the UI hides the rest",
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "roles": {
            "items": {
              "enum": [
                "viewer",
                "field_engineer",
                "rtmac_engineer",
                "drilling_engineer",
                "data_steward",
                "admin"
              ],
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "name",
          "roles",
          "permissions",
          "auth_mode"
        ],
        "title": "Me",
        "type": "object"
      },
...
      "ReviewAccept": {
        "additionalProperties": false,
        "description": "The proposed values are right: the target becomes verified.",
        "properties": {
          "action": {
            "const": "accept",
            "title": "Action",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "required": [
          "action"
        ],
        "title": "ReviewAccept",
        "type": "object"
      },
      "ReviewCorrect": {
        "additionalProperties": false,
        "description": "Replace some proposed values (same keys and canonical units as ``proposed``); the\ntarget is updated and verified, and the pair is kept for the evaluation gold set.",
        "properties": {
          "action": {
            "const": "correct",
            "title": "Action",
            "type": "string"
          },
          "fields": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "minProperties": 1,
            "title": "Fields",
            "type": "object"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "required": [
          "action",
          "fields"
        ],
        "title": "ReviewCorrect",
        "type": "object"
      },
      "ReviewItem": {
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "correction": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correction"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decided_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decided_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By"
          },
          "doc_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Type"
          },
          "document_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRef"
            },
            "title": "Evidence",
            "type": "array"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Field under review; null = the whole record",
            "title": "Field"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kind": {
            "enum": [
              "event",
              "mitigation",
              "casing",
              "cement",
              "mud",
              "alias",
              "other"
            ],
            "title": "Kind",
            "type": "string"
          },
          "page_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page No"
          },
          "proposed": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "description": "Extracted values, canonical units",
            "title": "Proposed",
            "type": "object"
          },
          "reason": {
            "description": "Why it needs review, e.g. 'confidence 0.62 < 0.75'",
            "title": "Reason",
            "type": "string"
          },
          "span_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Span Ids",
            "type": "array"
          },
          "status": {
            "enum": [
              "pending",
              "accepted",
              "corrected",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Row id in the table named by kind, if written",
            "title": "Target Id"
          },
          "well_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Well Id"
          },
          "well_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Well Name"
          }
        },
        "required": [
          "id",
          "kind",
          "target_id",
          "document_id",
          "filename",
          "doc_type",
          "well_id",
          "well_name",
          "page_no",
          "span_ids",
          "evidence",
          "field",
          "reason",
          "confidence",
          "proposed",
          "status",
          "decided_by",
          "decided_at",
          "correction",
          "created_at"
        ],
        "title": "ReviewItem",
        "type": "object"
      },
      "ReviewPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "status_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "description": "Items per status (all kinds)",
            "title": "Status Counts",
            "type": "object"
          }
        },
        "required": [
          "items",
          "next_cursor",
          "status_counts"
        ],
        "title": "ReviewPage",
        "type": "object"
      },
      "ReviewReject": {
        "additionalProperties": false,
        "description": "Not a real fact: the target is rejected (events get status 'rejected').",
        "properties": {
          "action": {
            "const": "reject",
            "title": "Action",
            "type": "string"
          },
          "reason": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "action",
          "reason"
        ],
        "title": "ReviewReject",
        "type": "object"
      },
...
      "TokenOut": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/Me"
          }
        },
        "required": [
          "access_token",
          "token_type",
          "expires_at",
          "user"
        ],
        "title": "TokenOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "eRTMAC-NWIS offset-well knowledge and decision support (SIH PS 121)",
    "title": "SMRITI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Wrong username or password"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Log in (jwt auth mode): a bearer token for one shift",
        "tags": [
          "system"
        ]
      }
    },
...
    "/api/v1/review-queue": {
      "get": {
        "description": "Items ordered by confidence (lowest first), then id. Pass ``status`` explicitly to\nsee decided items; ``status_counts`` always covers every status.",
        "operationId": "list_review_queue_api_v1_review_queue_get",
        "parameters": [
          {
            "description": "Default: pending only",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pending",
                    "accepted",
                    "corrected",
                    "rejected"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "pending",
              "description": "Default: pending only",
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "event",
                    "mitigation",
                    "casing",
                    "cement",
                    "mud",
                    "alias",
                    "other"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "document_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Document Id"
            }
          },
          {
            "description": "Page size",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Page size",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Opaque cursor from the previous page's next_cursor",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from the previous page's next_cursor",
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Low-confidence extractions awaiting review",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/review-queue/{item_id}": {
      "post": {
        "description": "Applies the decision to the target record (verified / corrected / rejected), records\n``decided_by``/``decided_at`` and returns the updated item. 409 if already decided.",
        "operationId": "review_item_api_v1_review_queue__item_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ReviewAccept"
                  },
                  {
                    "$ref": "#/components/schemas/ReviewCorrect"
                  },
                  {
                    "$ref": "#/components/schemas/ReviewReject"
                  }
                ],
                "title": "Decision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflicts with the current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept, correct or reject",
        "tags": [
          "review"
        ]
      }
    },
...
  }
}